抽象AddSound组件+优化音频信息录入
